Test now puts config into ~/snap/..., which works! \o/.
I had not realized that /home/jhartley, for a confined snap, is set to
, i.e. ~/snap/SNAPNAME/VERSION. So if my snap is
to read a config file, it has to be placed there. The test now
does this, and passes.

Also added a bit of color to tests, and removed the verbose
output now that it's passing.

Also, we don't need to grant personal-files read access to the
.ctags.d parent dir, that was a red herring.