Documentation and limitations for end-users.
README now directs readers to the user docs defined in
snap/snapcraft.yaml (or rendered online in the store.)

Snapcraft.yaml description now gives users some idea of
a) what doesn't work in this snap the same as in the apt.
b) Where to find detailed end-user docs for ctags.

Makefile gains new 'configure' target, which:
1. Adds a 'ctags' alias (useful if installing
from a local file, instead of the store, in which
case the store-approved alias is not automatically
provided.)
2. Connects the personal-files interface to allow
reading config from files in the ~/.ctags.d directory.
(At the moment I haven't asked the store to make this
auto-apply, but I think I can, and will.)

Tests now run using the "ctags" alias, instead of
"universal-ctags", in order to check the alias has
been applied properly.