Remove travis build badge, isn't building
Travis moved domain, switched to being more commercial,
plus started being kinda shifty. Somehow in the transition,
our original build has become a redirect to the new
domain, but the new domain returns a 404. So I don't
know if we're still building on Travis, and I'm not
inclined to try and get it working there.

Will consider an alternative CI at some point, but
the need for it is lessened somwhat by the automated
builds that snapcraft.io provide.